genai-testcase-generator/requirements.txt
# Core
Flask==2.3.2
gunicorn==20.1.0
pandas==2.2.2
openpyxl==3.1.2

# Vertex AI SDK
vertexai>=0.3.0

# Embeddings / RAG
faiss-cpu==1.7.4
numpy==1.26.0

# Streamlit frontend
streamlit==1.30.0
requests==2.31.0

# File parsing
python-docx==0.8.11
PyPDF2==3.0.1

# GCP clients (if using Secret Manager / Storage locally)
google-cloud-secret-manager==2.17.0
google-cloud-storage==2.9.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-api-core==2.11.0
